requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coverage==7.1.0
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.3
django-admin==2.0.2
django-allauth==0.51.0
django-crispy-forms==1.14.0
django_debug_toolbar==3.8.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-extensions==3.2.1
django-html-validator==0.5.1
django-six==1.0.4
django-webpack-loader==1.7.0
docutils==0.19
excel-base==1.0.4
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
isoweek==1.3.3
lxml==4.9.2
numpy==1.24.3
oauthlib==3.2.2
outcome==1.2.0
Pillow==9.3.0
pip==23.1.2
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
screen==1.0.1
selenium==4.7.2
sendgrid==6.9.7
setuptools==67.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
TimeConvert==3.0.13
trio==0.22.0
trio-websocket==0.9.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.13
wheel==0.40.0
wsproto==1.2.0
xlwt==1.3.0